Use new ID for live reporting period to ensure reset
The period was changed from seconds to milliseconds in b7b67f4. For
devices which had their previous setting stored in flash or EEPROM,
this led to publishing intervals of e.g. 1 ms instead of 1 s, which
may completely block the MCU trying to publish the reports.

By using a new ID, the previously stored value is ignored and the
default value is used instead.